+++ lib/src/i_channel.dart
@@ -14,14 +14,16 @@ class IChannel {
     this.child,
     this.buffer,
     this.assetsTexturePath,
+    this.texture,
     this.isSelf = false,
   }) : assert(
           !(!isSelf &&
               child != null &&
               buffer != null &&
-              assetsTexturePath != null),
-          'Only [isSelf] or [child] or [buffer] or [assetsTexturePath]'
-          ' must be given!',
+              assetsTexturePath != null &&
+              texture != null),
+          'Only [isSelf] or [child] or [buffer] or [assetsTexturePath] '
+          ' or [texture] must be given!',
         );
 
   /// the widget used by this [IChannel]
@@ -39,24 +41,30 @@ class IChannel {
   String? assetsTexturePath;
 
   /// the assets image if [assetsTexturePath] exists
-  ui.Image? assetsTexture;
+  ui.Image? texture;
 
   /// all textures loaded?
   bool isInited = false;
+
+  // ignore: use_setters_to_change_properties
+  void updateTexture(ui.Image image) => texture = image;
 
   /// eventually load textures
   Future<bool> init() async {
     if (isInited) return true;
 
     isInited = true;
-    // Load all the assets textures
+
+    if (texture != null) return isInited;
+
+    // Load the assets texture
     if (assetsTexturePath != null) {
       try {
         final assetImageByteData = await rootBundle.load(assetsTexturePath!);
         final codec = await ui.instantiateImageCodec(
           assetImageByteData.buffer.asUint8List(),
         );
-        assetsTexture = (await codec.getNextFrame()).image;
+        texture = (await codec.getNextFrame()).image;
       } catch (e) {
         debugPrint('Error loading assets image! $e');
         isInited = false;