+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.1.1
+- added `onShaderLoaded' callback to `ShaderBuffers`
+- fix the issue with the get of devicePixelRatio which caused a black widget.
+
 ## 1.1.0
 - fix Low Rendering Quality Issue #7
 - added `scaleRenderView` to `LayerBuffer` to compute the shader to a scaled resolution compared to the widget size.